Add some straight ports of ReplayBuffer and CartPoleEnv from python. Haven't tested anything yet. Need to either: a) test it all at once b) try incrementally using parts in python.
Will probably try a) first, and revert to b) when things (inevitably)
don't end up failing in ways I cannot debug.

+++ src/data_set.cc
@@ -49,7 +49,7 @@ void DataSet::ReadCSVFile(string dir, string file_name) {
   cout << "  max_price:" << data_set_metadata.max_price << endl;
 
   // the other lines contain the features for each car
-  for (int i = 2; i < lines.size(); ++i) {
+  for (size_t i = 2; i < lines.size(); ++i) {
     vector<float> features = ReadCSVLine(lines[i]);
     x_.insert(x_.end(), features.begin(), features.begin() + 3);
     y_.push_back(features[3]);
